Kalle/requirements/requirements.txt
discord.py
python-dotenv
yt-dlp
'discord[voice]'
wavelink